// web/src/types/RetentionQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EventNode } from "./EventNode";
import type { PropertyFilter } from "./PropertyFilter";
import type { RetentionPeriod } from "./RetentionPeriod";
import type { RetentionType } from "./RetentionType";

export type RetentionQuery = { 
/**
 * Event that places a person into a cohort.
 */
target: EventNode, 
/**
 * Event that counts as coming back.
 */
returning: EventNode, period: RetentionPeriod, 
/**
 * Number of cohorts and of periods per cohort.
 */
total_intervals: number, retention_type: RetentionType, properties: Array<PropertyFilter>, };